+++ examples/with-emailpassword/api-server.js
@@ -7,6 +7,7 @@ let Session = require("supertokens-node/recipe/session");
 let { verifySession } = require("supertokens-node/recipe/session/framework/express");
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let EmailPassword = require("supertokens-node/recipe/emailpassword");
+let Dashboard = require("supertokens-node/recipe/dashboard");
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
 const apiDomain = process.env.REACT_APP_API_URL || `http://localhost:${apiPort}`;
@@ -25,7 +26,13 @@ supertokens.init({
         apiDomain, // TODO: Change to your app's API domain
         websiteDomain, // TODO: Change to your app's website domain
     },
-    recipeList: [EmailPassword.init(), Session.init()],
+    recipeList: [
+        EmailPassword.init(),
+        Session.init(),
+        Dashboard.init({
+            apiKey: "supertokens_is_awesome",
+        }),
+    ],
 });
 
 const app = express();