+++ examples/with-thirdpartypasswordless/api-server/index.js
@@ -7,6 +7,7 @@ let Session = require("supertokens-node/recipe/session");
 let { verifySession } = require("supertokens-node/recipe/session/framework/express");
 let { middleware, errorHandler } = require("supertokens-node/framework/express");
 let ThirdPartyPasswordless = require("supertokens-node/recipe/thirdpartypasswordless");
+let Dashboard = require("supertokens-node/recipe/dashboard");
 require("dotenv").config();
 
 const apiPort = process.env.REACT_APP_API_PORT || 3001;
@@ -54,6 +55,9 @@ supertokens.init({
             flowType: "USER_INPUT_CODE_AND_MAGIC_LINK",
         }),
         Session.init(),
+        Dashboard.init({
+            apiKey: "supertokens_is_awesome",
+        }),
     ],
 });