+++ lib/ts/recipe/passwordless/index.ts
@@ -23,6 +23,7 @@ import { RecipeFunctionOptions, RecipeInterface } from "supertokens-web-js/recip
 import { PasswordlessFlowType, PasswordlessUser } from "supertokens-web-js/recipe/passwordless/types";
 import { getNormalisedUserContext } from "../../utils";
 import * as UtilFunctions from "./utils";
+import { PropsWithChildren } from "react";
 
 export default class Wrapper {
     static init(config: UserInput) {
@@ -191,7 +192,8 @@ export default class Wrapper {
 
     static PasswordlessAuth = PasswordlessAuth;
 
-    static SignInUp = (prop?: any) => Passwordless.getInstanceOrThrow().getFeatureComponent("signInUp", prop);
+    static SignInUp = (prop: PropsWithChildren<{ redirectOnSessionExists?: boolean; userContext?: any }> = {}) =>
+        Passwordless.getInstanceOrThrow().getFeatureComponent("signInUp", prop);
     static SignInUpTheme = SignInUpThemeWrapper;
 
     static LinkClicked = (prop?: any) =>