+++ test/with-typescript/src/App.tsx
@@ -101,6 +101,18 @@ function App() {
                                     </Auth>
                                 }
                             />
+                            <Route
+                                path="/no-redirection-sign-in"
+                                element={<EmailPassword.SignInAndUp redirectOnSessionExists={false} />}
+                            />
+                            <Route
+                                path="/no-redirection-sign-in-with-children"
+                                element={
+                                    <EmailPassword.SignInAndUp redirectOnSessionExists={false}>
+                                        <Home />
+                                    </EmailPassword.SignInAndUp>
+                                }
+                            />
                         </Routes>
                     </div>
                     <div className="footer">
@@ -1471,10 +1483,6 @@ ThirdPartyPasswordless.consumePasswordlessCode({
         preAPIHook: undefined,
     },
 });
-// @ts-expect-error
-ThirdPartyPasswordless.consumePasswordlessCode(undefined);
-// @ts-expect-error
-ThirdPartyPasswordless.consumePasswordlessCode();
 
 ThirdPartyPasswordless.createPasswordlessCode({
     email: "",